feat(DatePicker)!: replace v1 Cally-based date picker with pure-Stencil implementation
Removes the Cally-backed `bq-date-picker` (v1) and promotes the
pure-Stencil `bq-date-picker2` (v2) to be the default `bq-date-picker`.
The new implementation preserves the full v1 public API (props, events,
methods, slots, shadow parts, and CSS custom properties) and adds:

- `precision` prop — commit values at day / month / year granularity
- `initialView` prop — open the panel directly on days / months / years
- `bqViewChange` event — fired when the calendar switches views
- Native months and years views (no Cally dependency)
- 10 additional CSS custom properties for cell / header / nav theming
- 30+ additional shadow parts exposing the calendar internals

BREAKING CHANGES:

- The `tentative` attribute has been removed. The same tentative
  range-hover behavior is now handled as internal state; consumers
  should no longer set or read the `tentative` attribute.
- `bqChange.detail.value` is widened from `string` to `string | undefined`
  to align with the widened `value` prop; existing string-consumers
  continue to work but strict typings must be updated.
- `formatOptions` no longer has a hardcoded default. When omitted, the
  default is derived from `precision` (`day` → `{ day, month, year }`,
  `month` → `{ month, year }`, `year` → `{ year }`). Behavior at the
  default `day` precision is unchanged.
