+++ packages/beeq/src/components/date-picker/__tests__/labels.spec.ts
