+++ packages/beeq/src/components/date-picker/_storybook/bq-date-picker.mdx
@@ -4,23 +4,25 @@ import { ArgTypes, Title, Subtitle } from '@storybook/addon-docs/blocks';
   <div className="bq-doc__container">
     <Title>Date picker</Title>
 
-    The Date Picker is a intuitive UI element component allows users to select dates from a visual calendar interface, providing an intuitive way to input date information.
-    It offers a simple and efficient way to pick dates and/or time by minimizing errors and improving clarity.
+    Stencil-native rewrite of the Date Picker component (no Cally dependency). Same public API and wire format as `bq-date-picker` — additional controls include `initial-view` (open the calendar directly on months or years) and a fully Intl-driven header.
 
-    <Subtitle>Usage</Subtitle>
+    <Subtitle>Wire format</Subtitle>
 
-    - Use the Date Picker to allow users to select dates for scheduling appointments or events.
-    - Use the Date Picker to enable users to choose dates for travel bookings or reservations.
-    - Use the Date Picker to specify due dates or deadlines in project management tools.
-    - Use the Date Picker to filter records by date range in data analysis or reporting tools.
+    - `single`: `YYYY-MM-DD`
+    - `multi`:  `YYYY-MM-DD YYYY-MM-DD ...` (space separated)
+    - `range`:  `YYYY-MM-DD/YYYY-MM-DD`
 
-    <Subtitle>👍 When to use</Subtitle>
+    <Subtitle>Precision</Subtitle>
 
-    - When you need to collect precise date input, such as for booking systems, event scheduling, or project timelines.
-    - When providing a calendar view can enhance the user experience and reduce errors in date selection.
-    - When requiring users to select date ranges for filtering data, such as in reports or search functions.
-    - When simplifying the process of entering dates, reducing the likelihood of format errors, and improving overall data integrity.
-    - When offering users a convenient way to pick dates for reminders, deadlines, or other time-sensitive tasks.
+    Use the `precision` prop to change the granularity of the value:
+
+    - `day` (default): wire format is `YYYY-MM-DD`.
+    - `month`: wire format is `YYYY-MM`. The panel opens on the months view and clicking a month commits the value.
+    - `year`: wire format is `YYYY`. The panel opens on the years view and clicking a year commits the value.
+
+    All selection types (`single`, `multi`, `range`) work at every precision, e.g. a month range value looks like `YYYY-MM/YYYY-MM`.
+
+    When `formatOptions` is not provided, a sensible default is applied per precision (long month + year for `month`, year only for `year`). Consumer-provided `formatOptions` always take precedence.
 
     <Title>Properties</Title>
