--- packages/beeq/src/components/date-picker2/_storybook/bq-date-picker2.mdx
@@ -1,31 +1,0 @@
-import { ArgTypes, Title, Subtitle } from '@storybook/addon-docs/blocks';
-
-<div className="bq-doc__wrapper" data-theme="light">
-  <div className="bq-doc__container">
-    <Title>Date picker (v2)</Title>
-
-    Stencil-native rewrite of the Date Picker component (no Cally dependency). Same public API and wire format as `bq-date-picker` — additional controls include `initial-view` (open the calendar directly on months or years) and a fully Intl-driven header.
-
-    <Subtitle>Wire format</Subtitle>
-
-    - `single`: `YYYY-MM-DD`
-    - `multi`:  `YYYY-MM-DD YYYY-MM-DD ...` (space separated)
-    - `range`:  `YYYY-MM-DD/YYYY-MM-DD`
-
-    <Subtitle>Precision</Subtitle>
-
-    Use the `precision` prop to change the granularity of the value:
-
-    - `day` (default): wire format is `YYYY-MM-DD`.
-    - `month`: wire format is `YYYY-MM`. The panel opens on the months view and clicking a month commits the value.
-    - `year`: wire format is `YYYY`. The panel opens on the years view and clicking a year commits the value.
-
-    All selection types (`single`, `multi`, `range`) work at every precision, e.g. a month range value looks like `YYYY-MM/YYYY-MM`.
-
-    When `formatOptions` is not provided, a sensible default is applied per precision (long month + year for `month`, year only for `year`). Consumer-provided `formatOptions` always take precedence.
-
-    <Title>Properties</Title>
-
-    <ArgTypes of="bq-date-picker2" />
-  </div>
-</div>
