--- packages/beeq/src/components/date-picker2/scss/bq-date-picker2.variables.scss
@@ -1,78 +1,0 @@
-/* -------------------------------------------------------------------------- */
-/*                   Date picker v2 custom properties                         */
-/* -------------------------------------------------------------------------- */
-
-:host {
-  /**
-   * @prop --bq-date-picker2--background-color - Date picker background color
-   * @prop --bq-date-picker2--border-color - Date picker border color
-   * @prop --bq-date-picker2--border-color-disabled - Date picker border color when disabled
-   * @prop --bq-date-picker2--border-color-focus - Date picker border color on focus
-   * @prop --bq-date-picker2--border-radius - Date picker border radius
-   * @prop --bq-date-picker2--border-style - Date picker border style
-   * @prop --bq-date-picker2--border-width - Date picker border width
-   * @prop --bq-date-picker2--currentDate-border-color - Border color for today's date
-   * @prop --bq-date-picker2--currentDate-border-width - Border width for today's date
-   * @prop --bq-date-picker2--day-size - Size of a day cell
-   * @prop --bq-date-picker2--gap - Gap between input content and prefix/suffix
-   * @prop --bq-date-picker2--header-title-color - Header title text color
-   * @prop --bq-date-picker2--icon-size - Prefix/suffix/clear icon size
-   * @prop --bq-date-picker2--label-margin-bottom - Space below the label
-   * @prop --bq-date-picker2--label-text-color - Label text color
-   * @prop --bq-date-picker2--label-text-size - Label text size
-   * @prop --bq-date-picker2--nav-button-size - Header nav button size
-   * @prop --bq-date-picker2--padding-end - Input padding end
-   * @prop --bq-date-picker2--padding-start - Input padding start
-   * @prop --bq-date-picker2--paddingY - Input vertical padding
-   * @prop --bq-date-picker2--range-background-color - Range endpoint background
-   * @prop --bq-date-picker2--range-inner-background-color - Range inner background
-   * @prop --bq-date-picker2--row-gap - Vertical space between day-grid rows
-   * @prop --bq-date-picker2--text-color - Input text color
-   * @prop --bq-date-picker2--text-placeholder-color - Placeholder text color
-   * @prop --bq-date-picker2--text-size - Input text size
-   * @prop --bq-date-picker2--view-cell-background-hover - Hover background for day/month/year cells
-   * @prop --bq-date-picker2--view-cell-background-selected - Selected background for day/month/year cells
-   * @prop --bq-date-picker2--view-cell-height - Height of month/year cells
-   * @prop --bq-date-picker2--view-cell-radius - Border radius of day/month/year cells
-   * @prop --bq-date-picker2--view-cell-text-selected - Selected text color for day/month/year cells
-   */
-  --bq-date-picker2--background-color: var(--bq-ui--primary, #fff);
-  --bq-date-picker2--border-color: var(--bq-stroke--tertiary, #1c1d23);
-  --bq-date-picker2--border-color-disabled: var(--bq-stroke--secondary, #eaecf0);
-  --bq-date-picker2--border-color-focus: var(--bq-stroke--brand, #1570ef);
-  --bq-date-picker2--border-radius: var(--bq-radius--s, 0.5rem);
-  --bq-date-picker2--border-style: solid;
-  --bq-date-picker2--border-width: 1px;
-
-  --bq-date-picker2--currentDate-border-color: var(--bq-stroke--brand, #1570ef);
-  --bq-date-picker2--currentDate-border-width: 2px;
-
-  --bq-date-picker2--day-size: 2.25rem;         // 36px
-  --bq-date-picker2--gap: var(--bq-spacing-xs, 0.5rem);         // 8px
-  --bq-date-picker2--header-title-color: var(--bq-text--primary, #0f1728);
-  --bq-date-picker2--icon-size: 1.5rem;         // 24px
-
-  --bq-date-picker2--label-margin-bottom: var(--bq-spacing-xs, 0.5rem);
-  --bq-date-picker2--label-text-color: var(--bq-text--primary, #0f1728);
-  --bq-date-picker2--label-text-size: 0.875rem;
-
-  --bq-date-picker2--nav-button-size: 2rem;
-
-  --bq-date-picker2--padding-end: var(--bq-spacing-m, 1rem);
-  --bq-date-picker2--padding-start: var(--bq-spacing-m, 1rem);
-  --bq-date-picker2--paddingY: var(--bq-spacing-s, 0.75rem);
-
-  --bq-date-picker2--range-background-color: var(--bq-ui--brand, #1570ef);
-  --bq-date-picker2--range-inner-background-color: var(--bq-ui--brand-alt, #d1e9ff);
-  --bq-date-picker2--row-gap: 0.125rem;         // 2px — matches v1
-
-  --bq-date-picker2--text-color: var(--bq-text--primary, #0f1728);
-  --bq-date-picker2--text-placeholder-color: var(--bq-text--secondary, #667085);
-  --bq-date-picker2--text-size: 1rem;
-
-  --bq-date-picker2--view-cell-background-hover: var(--bq-ui--secondary-hover, #f2f4f7);
-  --bq-date-picker2--view-cell-background-selected: var(--bq-ui--brand, #1570ef);
-  --bq-date-picker2--view-cell-height: 2.75rem;
-  --bq-date-picker2--view-cell-radius: 0.5rem;
-  --bq-date-picker2--view-cell-text-selected: var(--bq-text--alt, #fff);
-}
