+++ packages/beeq/src/components/dropdown/readme.md
@@ -56,7 +56,6 @@ They provide an efficient way to save space and present a long list of options i
 ### Used by
 
  - [bq-date-picker](../date-picker)
- - [bq-date-picker2](../date-picker2)
  - [bq-select](../select)
 
 ### Depends on
@@ -68,7 +67,6 @@ They provide an efficient way to save space and present a long list of options i
 graph TD;
   bq-dropdown --> bq-panel
   bq-date-picker --> bq-dropdown
-  bq-date-picker2 --> bq-dropdown
   bq-select --> bq-dropdown
   style bq-dropdown fill:#f9f,stroke:#333,stroke-width:4px
 ```
